#22 Create Openstack config does not work properly since it requires list projects first
