+++ cce/cce.go
@@ -40,44 +40,9 @@ func GetKubeConfig(configParams KubeConfigParams) {
 	println(fmt.Sprintf("Successfully fetched and merge kube config for cce cluster %s.", configParams.ClusterName))
 }
 
-func GetProjectsInActiveCloud() {
-	projectsResponse := getProjectsFromServiceProvider()
-	var projects config.Projects
-	for _, project := range projectsResponse.Projects {
-		projects = append(projects, config.Project{
-			NameAndIdResource: config.NameAndIdResource{Name: project.Name, Id: project.Id},
-		})
-	}
-
-	config.UpdateProjects(projects)
-	println(fmt.Sprintf("Projects for active cloud:\n%s", strings.Join(projects.GetProjectNames(), ",\n")))
-}
-
-func getProjectsFromServiceProvider() (projectsResponse common.ProjectsResponse) {
-	cloud := config.GetActiveCloudConfig()
-	println(fmt.Sprintf("info: fetching projects for cloud %s", cloud.Domain.Name))
-
-	request := common.GetRequest(http.MethodGet, endpoints.IamProjects, nil)
-	request.Header.Add(headers.ContentType, headervalues.ApplicationJson)
-	request.Header.Add(xheaders.XAuthToken, cloud.UnscopedToken.Secret)
-
-	response := common.HttpClientMakeRequest(request)
-	bodyBytes := common.GetBodyBytesFromResponse(response)
-	projectsResponse = *common.DeserializeJsonForType[common.ProjectsResponse](bodyBytes)
-
-	return projectsResponse
-}
-
 func getClustersForProjectFromServiceProvider(projectName string) common.ClustersResponse {
 	clustersResponse := common.ClustersResponse{}
-	cloud := config.GetActiveCloudConfig()
-	project := cloud.Projects.FindProjectByName(projectName)
-	if project == nil {
-		GetProjectsInActiveCloud()
-		cloud = config.GetActiveCloudConfig()
-		verifiedProject := cloud.Projects.GetProjectByNameOrThrow(projectName)
-		project = &verifiedProject
-	}
+	project := config.GetActiveCloudConfig().Projects.GetProjectByNameOrThrow(projectName)
 
 	err := retry.Do(
 		func() error {
